requirements.txt
arrow==1.2.2
asgiref==3.5.2
black==22.3.0
blessed==1.19.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
croniter==1.3.5
cryptography==37.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.5
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
django-oauth-toolkit==2.0.0
django-picklefield==3.1
django-q==1.3.9
django-rest-knox==4.2.0
djangorestframework==3.13.1
flake8==4.0.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gspread==5.4.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.1
